+++ pkg/visor/visorconfig/v1.go
@@ -32,10 +32,10 @@ type V1 struct {
 
 	LogLevel             string                           `json:"log_level"`
 	LocalPath            string                           `json:"local_path"`
+	CustomDmsgHTTPPath   string                           `json:"custom_dmsg_http_path"`
 	StunServers          []string                         `json:"stun_servers"`
 	ShutdownTimeout      Duration                         `json:"shutdown_timeout,omitempty"`    // time value, examples: 10s, 1m, etc
 	RestartCheckDelay    Duration                         `json:"restart_check_delay,omitempty"` // time value, examples: 10s, 1m, etc
-	LogRotationInterval  Duration                         `json:"log_rotation_interval"`         // time value, examples: 10s, 1m, etc
 	IsPublic             bool                             `json:"is_public"`
 	PersistentTransports []transport.PersistentTransports `json:"persistent_transports"`
 
@@ -55,13 +55,15 @@ type Transport struct {
 	AddressResolver   string          `json:"address_resolver"`
 	PublicAutoconnect bool            `json:"public_autoconnect"`
 	TransportSetup    []cipher.PubKey `json:"transport_setup_nodes"`
+	LogStore          *LogStore       `json:"log_store"`
 }
 
 // LogStore configures a LogStore.
 type LogStore struct {
 	// Type defines the log store type. Valid values: file, memory.
-	Type     string `json:"type"`
-	Location string `json:"location"`
+	Type             string   `json:"type"`
+	Location         string   `json:"location"`
+	RotationInterval Duration `json:"rotation_interval"` // time value, examples: 10s, 1m, 1h etc
 }
 
 // Routing configures routing.
@@ -181,7 +183,7 @@ func (v1 *V1) GetPersistentTransports() ([]transport.PersistentTransports, error
 // UpdateLogRotationInterval updates log_rotation_interval in config
 func (v1 *V1) UpdateLogRotationInterval(d Duration) error {
 	v1.mu.Lock()
-	v1.LogRotationInterval = d
+	v1.Transport.LogStore.RotationInterval = d
 	v1.mu.Unlock()
 
 	return v1.flush(v1)
@@ -191,7 +193,7 @@ func (v1 *V1) UpdateLogRotationInterval(d Duration) error {
 func (v1 *V1) GetLogRotationInterval() (Duration, error) {
 	v1.mu.Lock()
 	defer v1.mu.Unlock()
-	return v1.LogRotationInterval, nil
+	return v1.Transport.LogStore.RotationInterval, nil
 }
 
 // UpdatePublicAutoconnect updates public_autoconnect in config